Another Hot Fix

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	long long int typeLong;
 	float floatType;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(charType));
-	printf("Size of an int: %ld byte(s)\n", sizeof(intType));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(longType));
-	printf("Size of a long long int: %ld byte(s)\n", sizeof(typeLong));
-	printf("Size of a float: %ld btye(s)\n", sizeof(floatType));
+	printf("Size of a char: %ld byte(s)\n", (unsigned long)sizeof(charType));
+	printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(intType));
+	printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(longType));
+	printf("Size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(typeLong));
+	printf("Size of a float: %ld btye(s)\n", (unsigned long)sizeof(floatType));
 
 	return (0);
-}	
+}
